Added Footer Added image on footer Added footer on Starter Page
