+++ next/styles/Footer.module.css
@@ -1,0 +1,6 @@
+.hand-job{
+  position: absolute;
+  width: 32rem;
+  top: 7rem;
+  left: 85rem;
+}
