+++ next/styles/Home.module.css
@@ -27,6 +27,6 @@
 .rocket-girl{
   position: absolute;
   width: 35rem;
-  top: 8rem;
+  top: 3rem;
   left: 75rem;
 }
