+++ next/styles/Tokenomics.module.css
@@ -1,7 +1,6 @@
 .third-page{
   background-image: url(../public/images/bg3.png);
   background-repeat: no-repeat;
-  background-size: contain;
-  height: 100vh;
-  
+  background-size: cover;
+  height: 80vh;
 }
